requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.2.1
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
fastapi==0.115.12
fonttools==4.51.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.2
openai==1.81.0
packaging==23.2
pandas==2.1.3
pillow==10.3.0
plotly==5.18.0
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.1.0
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.1
starlette==0.46.2
tenacity==8.2.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2023.3
uvicorn==0.34.2
